LayoutTests/media/modern-media-controls/play-pause-button/play-pause-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PlayPauseButton</code> class.");

const playButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("playButton.element.localName", "button");
shouldBeTrue('playButton.element.classList.contains("icon")');
shouldBeTrue('playButton.element.classList.contains("play-pause")');
shouldBeEqualToString("playButton.iconName", `${Icons.Play}`);
shouldBeFalse("playButton.playing");

const fullscreenPlayButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });
const iOSPlayButton = new PlayPauseButton({ layoutTraits: LayoutTraits.iOS });

const pauseButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS });
pauseButton.playing = true;
shouldBeTrue("pauseButton.playing");
shouldBeEqualToString("pauseButton.iconName", `${Icons.Pause}`);

const fullscreenPauseButton = new PlayPauseButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });
fullscreenPauseButton.playing = true;

const iOSPauseButton = new PlayPauseButton({ layoutTraits: LayoutTraits.iOS });
iOSPauseButton.playing = true;

shouldBecomeEqual('playButton.element.style.webkitMaskImage.includes("macOS/play@")', "true", () => {
    shouldBecomeEqual('pauseButton.element.style.webkitMaskImage.includes("macOS/pause@")', "true", () => {
        shouldBecomeEqual('iOSPlayButton.element.style.webkitMaskImage.includes("iOS/play@")', "true", () => {
            shouldBecomeEqual('iOSPauseButton.element.style.webkitMaskImage.includes("iOS/pause@")', "true", () => {
                shouldBecomeEqual('fullscreenPlayButton.element.style.webkitMaskImage.includes("macOS/play-fullscreen@")', "true", () => {
                    shouldBecomeEqual('fullscreenPauseButton.element.style.webkitMaskImage.includes("macOS/pause-fullscreen@")', "true", () => {
                        debug("");
                        finishJSTest();
                    });
                });
            });
        });
        
    });
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
