LayoutTests/media/modern-media-controls/play-pause-button/play-pause-button-expected.txt

Testing the PlayPauseButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS playButton.element.localName is "button"
PASS playButton.element.classList.contains("icon") is true
PASS playButton.element.classList.contains("play-pause") is true
PASS playButton.iconName is "play"
PASS playButton.playing is false
PASS pauseButton.playing is true
PASS pauseButton.iconName is "pause"
PASS playButton.element.style.webkitMaskImage.includes("macOS/play@") became true
PASS pauseButton.element.style.webkitMaskImage.includes("macOS/pause@") became true
PASS iOSPlayButton.element.style.webkitMaskImage.includes("iOS/play@") became true
PASS iOSPauseButton.element.style.webkitMaskImage.includes("iOS/pause@") became true
PASS fullscreenPlayButton.element.style.webkitMaskImage.includes("macOS/play-fullscreen@") became true
PASS fullscreenPauseButton.element.style.webkitMaskImage.includes("macOS/pause-fullscreen@") became true

PASS successfullyParsed is true

TEST COMPLETE
